apps/backend/pyproject.toml
[tool.poetry]
name = "spiritual-qa-backend"
version = "0.1.0"
description = "Backend for Spiritual Q&A Platform using RAG"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
argon2-cffi = "^23.1.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.28"
qdrant-client = "^1.8.0"
openai = "^1.13.3"
llama-index = "^0.10.18"
python-multipart = "^0.0.9"
alembic = "^1.13.1"
redis = {extras = ["hiredis"], version = "^5.0.0"}
email-validator = "^2.1.0"
python-dotenv = "^1.0.0"
pypdf = "^4.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"
ruff = "^0.3.0"
aiosqlite = "^0.20.0"
mutmut = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.coverage.run]
omit = [
    "app/api/test_router.py",
    "app/e2e_overrides.py",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
# addopts = "--cov=app --cov-report=term-missing --cov-fail-under=85"
markers = [
    "integration: marks tests requiring Docker services",
]

[tool.pyrefly]
project-includes = ["app", "tests", "scripts"]
python-version = "3.11"

[tool.mutmut]
paths_to_mutate = [
    # --- Tier 1: Pure Logic ---
    "app/trivial.py",
    "app/core/security.py",
    "app/rag/pdf_parser.py",
    "app/rag/chunker.py",
    "app/rag/embedder.py",
    "app/rag/prompts.py",
    # --- Tier 2: Services / Logic with Mock Dependencies ---
    "app/services/rate_limiter.py",
    "app/services/auth_service.py",
    "app/services/token_service.py",
    "app/services/conversation_service.py",
    "app/services/document_service.py",
    "app/services/rag_service.py",
    "app/services/token_cleanup.py",
    # --- Tier 3: Repositories & Store (Isolated via db_session/sqlite) ---
    "app/repositories/user_repo.py",
    "app/repositories/session_repo.py",
    "app/repositories/document_repo.py",
    "app/repositories/revoked_token_repo.py",
    "app/rag/qdrant_store.py",
    "app/rag/retriever.py",
    # --- Tier 4: Routers & Middleware (Isolated via Mock Clients) ---
    "app/api/admin_router.py",
    "app/api/auth_router.py",
    "app/api/chat_router.py",
    "app/middleware/body_limit.py",
    "app/middleware/csrf.py",
    "app/middleware/request_id.py",
    "app/middleware/security_headers.py"
]

also_copy = ["app/"]
debug = true
